MyBudgetPal/requirements.txt
Flask==3.1.1
Flask-Bcrypt==1.0.1
mysql-connector-python==9.3.0
gunicorn==23.0.0
DateTime==5.5
python-dotenv==1.1.0
google-genai==1.21.1
Flask-WTF==1.2.2
bleach==6.2.0
pyotp==2.9.0
qrcode[pil]==8.2
pytest==8.4.1
protobuf==6.31.1
qrcode==8.2
Requests==2.32.4
Werkzeug==3.1.3
